+++ HISTORY.md
@@ -6,6 +6,7 @@
 
 #### API
 
+- Make epsilon configurable in archives (#226)
 - **Backwards-incompatible:** Remove ribs.factory (#225)
 - **Backwards-incompatible:** Replaced `ImprovementEmitter`, `RandomDirectionEmitter`, and `OptimizingEmitter` with `EvolutionStrategyEmitter` (#220, #223)
 - Raise ValueError for incorrect array shapes in archive methods (#219)